fixed critical review comments (mapper config, loop safety, duplicates, interrupt handling)

+++ src/main/java/mate/academy/rickandmorty/dto/character/CreateCharacterRequestDto.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 public class CreateCharacterRequestDto {
     @JsonProperty("id")
-    private int externalId;
+    private Long externalId;
     private String name;
     private String status;
     private String gender;
